Convert old test cases to v2 schema (#37)
